Sales data
Sales data was analyzed by working in Excel and Google Looker Studio. The data is unfortunately without labels since I had issues finding labels on Google Looker Studio. The line graph shows sales data from January to December 2020. The column chart shows the same data but is grouped by month so we can see that the month with the highest sales is December. The pie chart shows the distribution of products by the quantity purchased, so we can see that AAA batteries were the most popular purchase. The donut chart shows the distribution of products by revenue, so we can see that selling Macbook Pros was the most lucrative for this company.